Split docbook-generator and rename docbook-adapter to docbook-renderer
 * Previously docbook-generator queried the SPARQL datastore to produce the document tree and generate the Docbook XML in one process. Now, the first stage is done by owl-doc (in owl-tools) and docbook-generator consumes its produced JSON.
 * Rename docbook-adapter to docbook-renderer`
