+++ docbook-tools/.classpath
@@ -6,5 +6,5 @@
 		</attributes>
 	</classpathentry>
 	<classpathentry kind="lib" path="gradle/wrapper/gradle-wrapper.jar"/>
-	<classpathentry kind="output" path="docbook-adapter/build/classes/java/main"/>
+	<classpathentry kind="output" path="docbook-renderer/build/classes/java/main"/>
 </classpath>
