+++ docbook-tools/docbook-renderer-gradle/build.gradle
@@ -5,5 +5,5 @@ apply plugin: 'java-library'
 
 dependencies {
 	implementation gradleApi()
-    implementation project (':docbook-adapter')
+    implementation project (':docbook-renderer')
 }
