+++ docbook-tools/docbook-renderer-gradle/src/main/resources/version.txt
