+++ docbook-tools/docbook-renderer/.settings/org.eclipse.jdt.core.prefs
