+++ docbook-tools/docbook-renderer/src/main/java/io/opencaesar/docbook/renderer/HTMLTransform.java
@@ -1,4 +1,4 @@
-package io.opencaesar.docbook.adapter;
+package io.opencaesar.docbook.renderer;
 
 import java.io.File;
