+++ docbook-tools/docbook-renderer/src/main/resources/log4j.xml
