+++ docbook-tools/docbook-renderer/src/main/resources/version.txt
