+++ docbook-tools/settings.gradle
@@ -1,5 +1,5 @@
-include 'docbook-adapter'
+include 'docbook-renderer'
 include 'docbook-generator'
 include 'docbook-stylesheets'
-include 'docbook-adapter-gradle'
+include 'docbook-renderer-gradle'
 include 'docbook-generator-gradle'
